+++ jenkins/jobs/dsl/infrastructure/infrastructure_pulumi_dashboard_job.groovy
@@ -58,6 +58,8 @@ pipelineJob(jobPath) {
         nonStoredPasswordParam('AWS_SESSION_TOKEN', 'AWS Session Token（オプション） - STS一時認証情報を使用する場合')
         
         // フィルタリングオプション
+        choiceParam('PROJECT_FILTER', projectFilterChoices, '''Pulumi プロジェクトの事前定義リスト（* で全件）'''.stripMargin())
+
         stringParam('PROJECT_FILTER', '', '''プロジェクト名フィルタ（Jenkins Agent 名も含む） - パターン入力で自由に絞り込み'''.stripMargin())
 
         stringParam('STACK_FILTER', '*', '''スタックフィルタ - 表示するスタックをフィルタリング